+++ source/sorting/helpers/partition.py
@@ -2,7 +2,7 @@ import math
 import random
 
 
-def partition(a: list[int], p: int, r: int) -> int:
+def partition(a: list[float], p: int, r: int) -> int:
     x = a[r - 1]
     i = p - 1
 
@@ -15,30 +15,31 @@ def partition(a: list[int], p: int, r: int) -> int:
     return i + 1
 
 
-def randomized_partition(a: list[int], p: int, r: int) -> int:
+def randomized_partition(a: list[float], p: int, r: int) -> int:
     i = random.randint(p, r - 1)
     a[i], a[r - 1] = a[r - 1], a[i]
     return partition(a, p, r)
 
 
-def partition_around(a: list[int], p: int, r: int, x: int) -> int:
+def partition_around(a: list[float], p: int, r: int, x: float) -> int:
     if x not in a:
         raise ValueError()
 
     i = 0
     while a[i] != x:
         i += 1
+    # i = a.index(x)
 
     a[i], a[r - 1] = a[r - 1], a[i]
     return partition(a, p, r)
 
 
-def good_partition(a: list[int], p: int, r: int) -> int:
+def good_partition(a: list[float], p: int, r: int) -> int:
     from source.sorting.helpers.select import select  # circular import
 
     n = r - p
     m = math.ceil(n / 5)
-    b: list[int] = []
+    b: list[float] = []
 
     for i in range(m):
         q = p + 5 * i
